give service calls unique callback IDs
Service calls made from Javascript now have a unique callback number
assigned. This number is passed to rosbridge, and then rosbridge passes
it back along with the answer.

This means that you can have multiple calls to the same service in
flight at the same time with different handlers.

The numerical value of the ID is never checked, so it should be possible
to pass anything JSON serializable as your callback ID, at the expense
of sending that data over the network and back.